backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
websockets==15.0.1
wheel==0.45.1
numpy==2.2.5
pillow==11.2.1
pygltflib==1.16.4
